+++ src/popup/pages/SignTransaction/SignView/index.js
@@ -63,6 +63,7 @@ import {
   parsedZekoFee,
 } from "../../../../utils/utils";
 import CountdownTimer from "../../../component/CountdownTimer";
+import { requestLedgerSignMessage } from "../../../../utils/ledger";
 import { TypeRowInfo } from "../TypeRowInfo";
 import styles from "./index.module.scss";
 
@@ -95,6 +96,8 @@ const COMMON_TRANSACTION_ACTION = [
 const Ledger_support_action = [
   DAppActions.mina_sendPayment,
   DAppActions.mina_sendStakeDelegation,
+  DAppActions.mina_signMessage,
+  DAppActions.mina_sign_JsonMessage,
 ];
 
 const StyledJsonView = styled.div`
@@ -514,6 +517,30 @@ const SignView = ({
           postRes = await sendStakeTx(payload, {
             rawSignature: signature,
           }).catch((error) => error);
+        } else if (
+          sendAction === DAppActions.mina_signMessage ||
+          sendAction === DAppActions.mina_sign_JsonMessage
+        ) {
+          let nextMsg = params.message;
+          if (sendAction === DAppActions.mina_sign_JsonMessage) {
+            nextMsg = JSON.stringify(params.message);
+          }
+          signResult = await requestLedgerSignMessage(
+            nextLedgerApp,
+            nextMsg,
+            currentAccount.hdPath
+          );
+          const { signature, signedMessage, error } = signResult;
+          if (error) {
+            setConfirmModalStatus(false);
+            Toast.info(error.message);
+            return;
+          }
+          postRes = {
+            data: signedMessage,
+            publicKey: params.fromAddress,
+            signature: signature,
+          };
         }
 
         setConfirmModalStatus(false);
@@ -669,10 +696,6 @@ const SignView = ({
           break;
         case DAppActions.mina_createNullifier:
           resultAction = DAPP_ACTION_CREATE_NULLIFIER;
-          break;
-        case DAppActions.mina_signMessage:
-        case DAppActions.mina_sign_JsonMessage:
-          resultAction = DAPP_ACTION_SIGN_MESSAGE;
           break;
         case DAppActions.mina_storePrivateCredential:
           resultAction = DAPP_ACTION_STORE_CREDENTIAL;